NO JIRA. Added support for zipped bags

+++ src/main/java/nl/knaw/dans/easy/sword2examples/ContinuedDeposit.java
@@ -44,11 +44,15 @@ public class ContinuedDeposit {
         depositPackage(new File(bagFileName), colIri, uid, pw, chunkSize);
     }
 
-    public static URI depositPackage(File bagDir, IRI colIri, String uid, String pw, int chunkSize) throws Exception {
-        // 0. Zip the bagDir
-        File zipFile = new File(bagDir.getAbsolutePath() + ".zip");
-        zipFile.delete();
-        Common.zipDirectory(bagDir, zipFile);
+    public static URI depositPackage(File bag, IRI colIri, String uid, String pw, int chunkSize) throws Exception {
+        File zipFile = null;
+        if (bag.isDirectory()) {
+            zipFile = new File(bag.getAbsolutePath() + ".zip");
+            zipFile.delete();
+            Common.zipDirectory(bag, zipFile);
+        } else {
+            zipFile = bag;
+        }
 
         // 1. Set up stream for calculating MD5
         FileInputStream fis = new FileInputStream(zipFile);
@@ -75,7 +79,7 @@ public class ContinuedDeposit {
         Link seIriLink = receipt.getLink("edit");
         URI seIri = seIriLink.getHref().toURI();
 
-        int remaining = (int) zipFile.length() - chunkSize;
+        long remaining = zipFile.length() - chunkSize;
         int count = 2;
         while (remaining > 0) {
             System.out.print(String.format("POST-ing chunk of %d bytes to SE-IRI (remaining: %d) ... ", chunkSize, remaining));
